pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
show_missing = true
fail_under = 95
exclude_also = [
  "pragma: nocover",
  "pragma: no cover",
  "import",
  "if self.debug",
  "raise AssertionError",
  "raise NotImplementedError",
  "raise RuntimeError",
  "raise TypeError",
  "def __repr__",
  "def __str__",
  "assert False",
  "pass,"
]

[tool.coverage.run]
source = ["src/*"]
branch = "true"
omit = [
  "env/*",
  "*/asgi.py",
  "*/wsgi.py",
  "*/migrations/*.py",
  "*/settings.py",
  "*/manage.py",
  "*/apps.py",
  "*/setup.py",
  "/pythonenv/**",
  "*/*.pyx",
  "*/*.pxd",
  "*/llm/service/placeholder.py"
]
disable_warnings = ["module-not-measured"]
concurrency = ["multiprocessing"]
parallel = true

[tool.mypy]
exclude = [
  "env/",
  ".venv/",
  "local",
  "build"
]
python_version = 3.13
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = false
disallow_untyped_defs = true
no_implicit_optional = true
strict = true
follow_imports = "silent"
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.poetry]
name = "albatross"
version = "0.0.4"
description = ""
authors = [
  "MPB Development <dev@mpb.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
google-generativeai = "^0.8.4"
mysql-connector-python = "^9.2.0"
grpcio = "*"
grpcio-health-checking = "*"
grpcio-tools = "*"
toml-sort = "^0.24.2"
openai = "^1.69.0"
google-genai = "^1.8.0"
piexif = "^1.1.3"
sqlalchemy = "^2.0.40"
mypy = "*"
pytest = "*"
pytest-bdd = "*"
pytest-cov = "*"
pytest-xdist = "*"
setuptools = "*"
pytest-freezegun = "*"
ruff = "*"
certifi = "*"
charset-normalizer = "*"
click = "*"
Flask = "*"
idna = "*"
importlib-metadata = "*"
itsdangerous = "*"
Jinja2 = "*"
MarkupSafe = "*"
mypy-extensions = "*"
pathspec = "*"
platformdirs = "*"
requests = "*"
tomli = "*"
typing_extensions = "*"
urllib3 = "*"
Werkzeug = "*"
zipp = "*"
pipdeptree = "^2.25.0"
black = "^25.1.0"
flask-restx = "^1.3.0"
google-cloud-aiplatform = "^1.89.0"
types-requests = "^2.32.0.20250328"
python-dotenv = "^1.1.0"

[[tool.poetry.source]]
name = "mpb-core"
url = "https://mpb:26W13PyPiRepo@pypi.mpb.com/simple/"
priority = "supplemental"

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
addopts = "--cov=src --cov-report=term-missing --ignore=common/tests"

[tool.ruff]
line-length = 88
exclude = [
  ".git",
  ".venv",
  "env",
  "tests/fixtures"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "Q",
  "TID252"
]

[tool.ruff.lint.isort]
known-first-party = [
  "albatross",
  "tests"
]
